Merged revisions 916 via svnmerge from https://origsvn.digium.com/svn/asterisk-gui/trunk
........
r916 | bkruse | 2007-05-10 18:44:51 -0500 (Thu, 10 May 2007) | 1 line

Now, you can use the options page in the gui to make agentlogin and agentcallback login extensions availible.....The only thing left to do is to UPDATE the extension, instead of as it is now, where it just adds another extension if you change the number. I am sure there is a way to do this, pari can help me. Anyways, the basis of it is layed out now.
........