Merged revisions 1097 via svnmerge from https://origsvn.digium.com/svn/asterisk-gui/trunk
........
r1097 | bkruse | 2007-06-18 19:51:09 -0500 (Mon, 18 Jun 2007) | 1 line

Wow, this will fix alot of problems. Instead of doing the ghetto way of naming a variable then executing a system command, we just call a local channel that answers and hangs up, and the oppisite application: is system. I am not sure if this makes things faster, but now nothing is hanging and all the system calls are ALOT smoother, and you can actually tell from the responseText if the command was successfull (instead of it always returning false like it used to).  I think this will help alot of things. Check it over for me pari. The last problem had to do with us reloading WHILE we were calling for a config file or system call. Since we do not use callbacks there, it would execute right after executing the reload. But its fixed now, yay
........