BugFix: http://bugs.digium.com/view.php?id=10366
When creating a new call group, the number of seconds to ring the group is 20 by default.
When you save the call group, the number of seconds is set to "undefined". You can edit 
the call group and then set the ring seconds, but the issue is with adding the 
group the first time.