Merged revisions 1332 via svnmerge from https://origsvn.digium.com/svn/asterisk-gui/branches/1.4
........
r1332 | bkruse | 2007-08-14 16:56:16 -0500 (Tue, 14 Aug 2007) | 1 line

Initial tool (ztscan) to scan for digital cards. Outputs to file /etc/asterisk/ztscan.conf to be read in and parsed by the GUI. This is still in BETA, but works pretty well. Pari and I will be working on the GUI aspect of this throughout the week. The tool uses /dev/zap/ctl (ioctl) to get span information for each span. Todo: use ZT_GET_PARAMS instead of ZT_SPAN_INFO to get the lbo/framing/coding etc. ztcfg does this, but that is because its setting the arguments with ZT_SET_PARAM, not retreiving them. Other than that, works great. The config file will tell the gui continue=yes or continue=no depending on whether or not cards were found in the system. The tool has the ability to modprobe for drivers (but does not currently do so) since that is up to the OS. The config file also has the current alarm of the span, the total spans inside the general section (for the gui, if t.fieldbyname['totalspans'] < 0, return false, etc), Description of the card, Name of the card, how many channels it has, and how many channels are currently configured.
........