Alot more changes for digital, moving in the right direction.
My idea now is to do all the zaptel configuration in digital.html,
including adding the trunks to zapata.conf.
Then in trunks.html, it will be like analog configuration, configuring
each line (or span) to an individual trunk. This method should not
take long to implement at all, yay.