Merged revisions 1888 via svnmerge from https://origsvn.digium.com/svn/asterisk-gui/branches/1.4
........
r1888 | bkruse | 2007-12-04 09:26:51 -0600 (Tue, 04 Dec 2007) | 13 lines

Removing ztscan/zapscan as they are in
zaptel now. (merged as one tool called ztscan.c)

Note: We can probably remove (not sure) the 
configure script, as it may no longer be needed.
(As we do not have to check for zaptel any longer)

I am glad we did it this way, and thanks to Kevin
for making ztscan awesome and 'fully functional' :)

It is now in
zaptel/branches/1.4/ztscan.c

........