I made a comment in the graphs page (if anyone actually uncomments and accesses it)
Took the references out of the Makefile.

Fixed an example in astman.js referencing it (just for my peace of mind)

Removed the graphs.sh script, just to make sure.

I left everything in the tooltips, incase we add
it back we can just re-add the script, makefile,
and uncomment the embed in graphs.html