Fix for : Span numbering can confuse the user
Description: 

ztscan numbers the cards and spans with [x] this is the number that is assigned to the spans in the Service Providers tab. The problem is the spans 
show up as 1,2 in hardware configuration page.

Example
In ZTscan
TDM800p [1]
TE220p [2]
TE220p [3]
so, the spans will be numbered 2and 3 in the service providers tab but will show up as 1 and 2 in the setup hardware tab. 

Reported By: Sean Pimental